Moved the popped out window into its own compiled project as a sibling project of the main.
Working with react context and window.open was creating a working setup for most things, but not for drag and drop.
This resolves the problem with drag and drop as well as all styling issues without having to set up another repository or repeating myself.
The two projects communicate with each other via sockets as I haven't found a reliable way to allow for cross communication that works in both dev and production.
This change also spins up two webpack-dev-server instances, one for each project, so that they can be served independently.